requirements.txt
flake8==5.0.4
isort==5.10.1
Pillow==9.4.0